@@ -1,122 +1,117 @@
 ﻿using Microsoft.AspNetCore.Components;
-using System;
-using System.IO;
-using System.Net.Http;
-using System.Threading.Tasks;
 using System.Web;
 
-namespace Tavenem.Blazor.ImageEditor.Sample.Client.Pages
+namespace Tavenem.Blazor.ImageEditor.Sample.Client.Pages;
+
+public partial class Index
 {
-    public partial class Index
+    [Inject] private HttpClient? HttpClient { get; set; }
+
+    private ImageEditor? ImageEditor { get; set; }
+
+    private protected bool IsLoading { get; set; }
+
+    private protected string LoadText { get; set; } = "Loading";
+
+    protected override async Task OnAfterRenderAsync(bool firstRender)
     {
-        [Inject] private HttpClient? HttpClient { get; set; }
-
-        private ImageEditor? ImageEditor { get; set; }
-
-        private protected bool IsLoading { get; set; }
-
-        private protected string LoadText { get; set; } = "Loading";
-
-        protected override async Task OnAfterRenderAsync(bool firstRender)
+        if (!firstRender || HttpClient is null)
         {
-            if (!firstRender || HttpClient is null)
-            {
-                return;
-            }
+            return;
+        }
 
-            LoadText = "Loading";
-            IsLoading = true;
-            StateHasChanged();
-            await Task.Delay(1).ConfigureAwait(false);
+        LoadText = "Loading";
+        IsLoading = true;
+        StateHasChanged();
+        await Task.Delay(1).ConfigureAwait(false);
 
-            var host = await HttpClient.GetStringAsync("/upload").ConfigureAwait(false);
-            if (string.IsNullOrEmpty(host))
-            {
-                IsLoading = false;
-                StateHasChanged();
-                return;
-            }
-
-            if (ImageEditor is not null)
-            {
-                if (await FileExistsAsync("editorimage.png").ConfigureAwait(false))
-                {
-                    await ImageEditor.LoadImageAsync(Path.Combine(host, "editorimage.png")).ConfigureAwait(false);
-                }
-                if (await FileExistsAsync("editorjson.json").ConfigureAwait(false))
-                {
-                    var json = await HttpClient.GetStringAsync(Path.Combine(host, "editorjson.json")).ConfigureAwait(false);
-                    await ImageEditor.LoadJSONAsync(json).ConfigureAwait(false);
-                }
-            }
-
+        var host = await HttpClient.GetStringAsync("/upload").ConfigureAwait(false);
+        if (string.IsNullOrEmpty(host))
+        {
             IsLoading = false;
             StateHasChanged();
+            return;
         }
 
-        private protected async Task<bool> FileExistsAsync(string name)
+        if (ImageEditor is not null)
         {
-            if (HttpClient is null)
+            if (await FileExistsAsync("editorimage.png").ConfigureAwait(false))
             {
-                return false;
+                await ImageEditor.LoadImageAsync(Path.Combine(host, "editorimage.png")).ConfigureAwait(false);
             }
-            try
+            if (await FileExistsAsync("editorjson.json").ConfigureAwait(false))
             {
-                var response = await HttpClient.GetAsync(
-                    $"/upload/exists?name={HttpUtility.UrlEncode(name)}")
-                    .ConfigureAwait(false);
-                return response.IsSuccessStatusCode;
-            }
-            catch
-            {
-                return false;
+                var json = await HttpClient.GetStringAsync(Path.Combine(host, "editorjson.json")).ConfigureAwait(false);
+                await ImageEditor.LoadJSONAsync(json).ConfigureAwait(false);
             }
         }
 
-        private async Task GetObjUrlAsync(string? url)
+        IsLoading = false;
+        StateHasChanged();
+    }
+
+    private protected async Task<bool> FileExistsAsync(string name)
+    {
+        if (HttpClient is null)
         {
-            if (HttpClient is null
-                || string.IsNullOrEmpty(url))
-            {
-                return;
-            }
+            return false;
+        }
+        try
+        {
+            var response = await HttpClient.GetAsync(
+                $"/upload/exists?name={HttpUtility.UrlEncode(name)}")
+                .ConfigureAwait(false);
+            return response.IsSuccessStatusCode;
+        }
+        catch
+        {
+            return false;
+        }
+    }
 
-            try
-            {
-                var bytes = await HttpClient
-                    .GetByteArrayAsync(url)
-                    .ConfigureAwait(false);
-
-                _ = await HttpClient.PostAsync(
-                    "/upload?name=editorimage.png",
-                    new ByteArrayContent(bytes))
-                    .ConfigureAwait(false);
-            }
-            catch (Exception ex)
-            {
-                Console.WriteLine($"Error uploading image: {ex.Message}");
-            }
+    private async Task GetObjUrlAsync(string? url)
+    {
+        if (HttpClient is null
+            || string.IsNullOrEmpty(url))
+        {
+            return;
         }
 
-        private async Task SaveJsonAsync(string? json)
+        try
         {
-            if (HttpClient is null
-                || string.IsNullOrEmpty(json))
-            {
-                return;
-            }
+            var bytes = await HttpClient
+                .GetByteArrayAsync(url)
+                .ConfigureAwait(false);
 
-            try
-            {
-                _ = await HttpClient.PostAsync(
-                    "/upload?name=editorjson.json",
-                    new StringContent(json))
-                    .ConfigureAwait(false);
-            }
-            catch (Exception ex)
-            {
-                Console.WriteLine($"Error uploading json: {ex.Message}");
-            }
+            _ = await HttpClient.PostAsync(
+                "/upload?name=editorimage.png",
+                new ByteArrayContent(bytes))
+                .ConfigureAwait(false);
+        }
+        catch (Exception ex)
+        {
+            Console.WriteLine($"Error uploading image: {ex.Message}");
+        }
+    }
+
+    private async Task SaveJsonAsync(string? json)
+    {
+        if (HttpClient is null
+            || string.IsNullOrEmpty(json))
+        {
+            return;
+        }
+
+        try
+        {
+            _ = await HttpClient.PostAsync(
+                "/upload?name=editorjson.json",
+                new StringContent(json))
+                .ConfigureAwait(false);
+        }
+        catch (Exception ex)
+        {
+            Console.WriteLine($"Error uploading json: {ex.Message}");
         }
     }
 }
