@@ -1,43 +1,42 @@
-﻿namespace Tavenem.Blazor.ImageEditor
+﻿namespace Tavenem.Blazor.ImageEditor;
+
+/// <summary>
+/// A drawing mode supported by the image editor.
+/// </summary>
+public enum DrawingMode
 {
     /// <summary>
-    /// A drawing mode supported by the image editor.
+    /// Ellipses, including circles.
     /// </summary>
-    public enum DrawingMode
-    {
-        /// <summary>
-        /// Ellipses, including circles.
-        /// </summary>
-        Ellipse,
+    Ellipse = 0,
 
-        /// <summary>
-        /// Straight lines.
-        /// </summary>
-        Line,
+    /// <summary>
+    /// Straight lines.
+    /// </summary>
+    Line = 1,
 
-        /// <summary>
-        /// Rectangles, including squares.
-        /// </summary>
-        Rectangle,
+    /// <summary>
+    /// Rectangles, including squares.
+    /// </summary>
+    Rectangle = 2,
 
-        /// <summary>
-        /// Curved paths (quadratic curves).
-        /// </summary>
-        Path,
+    /// <summary>
+    /// Curved paths (quadratic curves).
+    /// </summary>
+    Path = 3,
 
-        /// <summary>
-        /// Text.
-        /// </summary>
-        Text,
+    /// <summary>
+    /// Text.
+    /// </summary>
+    Text = 4,
 
-        /// <summary>
-        /// Triangles.
-        /// </summary>
-        Triangle,
+    /// <summary>
+    /// Triangles.
+    /// </summary>
+    Triangle = 5,
 
-        /// <summary>
-        /// Free drawing mode.
-        /// </summary>
-        Free,
-    }
+    /// <summary>
+    /// Free drawing mode.
+    /// </summary>
+    Free = 6,
 }
