@@ -1,23 +1,11 @@
 using Microsoft.AspNetCore.Components.WebAssembly.Hosting;
-using Microsoft.Extensions.DependencyInjection;
-using System;
-using System.Net.Http;
-using System.Threading.Tasks;
+using Tavenem.Blazor.ImageEditor.Sample.Client;
 
-namespace Tavenem.Blazor.ImageEditor.Sample.Client
-{
-    public static class Program
-    {
-        public static async Task Main(string[] args)
-        {
-            var builder = WebAssemblyHostBuilder.CreateDefault(args);
-            builder.RootComponents.Add<App>("#app");
+var builder = WebAssemblyHostBuilder.CreateDefault(args);
+builder.RootComponents.Add<App>("#app");
 
-            builder.Services.AddScoped(sp => new HttpClient { BaseAddress = new Uri(builder.HostEnvironment.BaseAddress) });
+builder.Services.AddScoped(sp => new HttpClient { BaseAddress = new Uri(builder.HostEnvironment.BaseAddress) });
 
-            builder.Services.AddImageEditor();
+builder.Services.AddImageEditor();
 
-            await builder.Build().RunAsync().ConfigureAwait(false);
-        }
-    }
-}
+await builder.Build().RunAsync().ConfigureAwait(false);
