@@ -1,2 +1,2 @@
-class t{constructor(){this.originX=0,this.originY=0,this.drawingMode=0}make(t,i,s,e,a,n){if(this.originX=t,this.originY=i,n){const a=Math.max(Math.abs(t-s),Math.abs(i-e));s=t>s?t-a:t+a,e=i>e?i-a:i+a}return new fabric.Ellipse({left:s<t?s:t,top:e<i?e:i,rx:Math.abs(s-t)/2,ry:Math.abs(s-t)/2,...a})}resize(t,i,s,e,a,n){if(t instanceof fabric.Ellipse){if(n){const t=Math.max(Math.abs(this.originX-i),Math.abs(this.originY-s));i=this.originX>i?this.originX-t:this.originX+t,s=this.originY>s?this.originY-t:this.originY+t}t.set({originX:this.originX>i?"right":"left",originY:this.originY>s?"bottom":"top",rx:Math.abs(this.originX-i)/2,ry:Math.abs(this.originY-s)/2}).setCoords()}return t}}class i{constructor(){this.drawingMode=1}make(t,i,s,e,a,n){if(n){const a=Math.abs(t-s),n=Math.abs(i-e),r=Math.max(a,n);Math.abs(a-n)<=3?(s=t>s?t-r:t+r,e=i>e?i-r:i+r):a>n?(s=t>s?t-r:t+r,e=i):(s=t,e=i>e?i-r:i+r)}return new fabric.Line([t,i,s,e],a)}resize(t,i,s,e,a,n){if(t instanceof fabric.Line){if(n){const e=Math.abs((t.x1??0)-i),a=Math.abs((t.y1??0)-s),n=Math.max(e,a);Math.abs(e-a)<=3?(i=(t.x1??0)>i?(t.x1??0)-n:(t.x1??0)+n,s=(t.y1??0)>s?(t.y1??0)-n:(t.y1??0)+n):e>a?(i=(t.x1??0)>i?(t.x1??0)-n:(t.x1??0)+n,s=t.y1??0):(i=t.x1??0,s=(t.y1??0)>s?(t.y1??0)-n:(t.y1??0)+n)}t.set({x2:i,y2:s}).setCoords()}return t}}class s{constructor(){this.drawingMode=1}addPoint(t,i,s){const e=t.path,a=e.length+1;e.push(["Q",i,s,i,s]);const n=t.getCenterPoint(),r=new fabric.Path(e);t.set({path:r.path,width:r.width,height:r.height,dirty:!0,pathOffset:r.pathOffset}).setCoords(),t.setPositionByOrigin(n,"center","center"),this.addControlPoint(t,a-1)}make(t,i,s,e,a,n){const r=[[],[]];if(n){const a=Math.abs(t-s),n=Math.abs(i-e),r=Math.max(a,n);Math.abs(a-n)<=3?(s=t>s?t-r:t+r,e=i>e?i-r:i+r):a>n?(s=t>s?t-r:t+r,e=i):(s=t,e=i>e?i-r:i+r)}r[0][0]="M",r[0][1]=t,r[0][2]=i,r[1][0]="Q",r[1][1]=s,r[1][2]=e,r[1][3]=s,r[1][4]=e;const o=new fabric.Path(r,a);return o.controls={...fabric.Object.prototype.controls},o.controls.csc=new fabric.Control({positionHandler:this.curvePositionHandler,actionHandler:this.curveActionHandler,actionName:"modifyCurve",pointIndex:0}),this.addControlPoint(o,1),o}resize(t,i,s,e,a,n){if(t instanceof fabric.Path&&t.path){const e=t.path,a=e.length-1;if(n&&1===a){const t=Math.abs(e[0][1]-i),a=Math.abs(e[0][2]-s),n=Math.max(t,a);Math.abs(t-a)<=3?(i=e[0][1]>i?e[0][1]-n:e[0][1]+n,s=e[0][2]>s?e[0][2]-n:e[0][2]+n):t>a?(i=e[0][1]>i?e[0][1]-n:e[0][1]+n,s=e[0][2]):(i=e[0][1],s=e[0][2]>s?e[0][2]-n:e[0][2]+n)}const r=e;r[a][1]=i,r[a][2]=s,r[a][3]=i,r[a][4]=s;const o=t.getCenterPoint(),h=new fabric.Path(r);t.set({path:h.path,width:h.width,height:h.height,dirty:!0,pathOffset:h.pathOffset}),1===a&&t.set({left:i<e[0][1]?i:e[0][1],top:s<e[0][2]?s:e[0][2]}),t.setCoords(),1!==a&&t.setPositionByOrigin(o,"center","center")}return t}addControlPoint(t,i){t.controls[`cqc${i}`]=new fabric.Control({positionHandler:this.curvePositionHandler,actionHandler:this.curveActionHandler,actionName:"modifyCurve",pointIndex:i,pointIsCurve:!0}),t.controls[`cpc${i}`]=new fabric.Control({positionHandler:this.curvePositionHandler,actionHandler:this.curveActionHandler,actionName:"modifyCurve",pointIndex:i})}curveActionHandler(t,i,s,e){if(!(i.target instanceof fabric.Path))return!1;const a=i.target,n=a.path,r=a.controls[a.__corner],o=a._getNonTransformedDimensions(),h=a._getTransformedDimensions(),d=a.strokeWidth??1,f=a.strokeUniform?d:d*(a.scaleX??1),l=(a.strokeUniform||a.scaleY,(o.x-d)/(h.x-f)),b=(o.y-d)/(h.y-f),p=a.toLocalPoint(new fabric.Point(s,e),"center","center"),v=new fabric.Point(p.x*l+a.pathOffset.x,p.y*b+a.pathOffset.y);0===r.pointIndex?(n[0][1]=v.x,n[0][2]=v.y):r.pointIsCurve?c&&c.shift?(n[r.pointIndex][1]=n[r.pointIndex][3],n[r.pointIndex][2]=n[r.pointIndex][4]):(n[r.pointIndex][1]=v.x,n[r.pointIndex][2]=v.y):(n[r.pointIndex][1]===n[r.pointIndex][3]&&n[r.pointIndex][2]===n[r.pointIndex][4]&&(n[r.pointIndex][1]=v.x,n[r.pointIndex][2]=v.y),n[r.pointIndex][3]=v.x,n[r.pointIndex][4]=v.y);const u=a.getCenterPoint(),g=new fabric.Path(n);return a.set({path:g.path,width:g.width,height:g.height,dirty:!0,pathOffset:g.pathOffset}).setCoords(),a.setPositionByOrigin(u,"center","center"),!0}curvePositionHandler(t,i,s){if(!(s instanceof fabric.Path))return new fabric.Point(0,0);const e=s.path;let a,n;if(0===this.pointIndex)a=e[0][1]-s.pathOffset.x,n=e[0][2]-s.pathOffset.y;else if(this.pointIsCurve)if(e[this.pointIndex][1]===e[this.pointIndex][3]&&e[this.pointIndex][2]===e[this.pointIndex][4]){const t=1===this.pointIndex?1:3,i=1===this.pointIndex?2:4;a=(e[this.pointIndex-1][t]+e[this.pointIndex][3])/2-s.pathOffset.x,n=(e[this.pointIndex-1][i]+e[this.pointIndex][4])/2-s.pathOffset.y}else a=e[this.pointIndex][1]-s.pathOffset.x,n=e[this.pointIndex][2]-s.pathOffset.y;else a=e[this.pointIndex][3]-s.pathOffset.x,n=e[this.pointIndex][4]-s.pathOffset.y;return s.canvas&&s.canvas.viewportTransform?fabric.util.transformPoint(new fabric.Point(a,n),fabric.util.multiplyTransformMatrices(s.canvas.viewportTransform,s.calcTransformMatrix())):fabric.util.transformPoint(new fabric.Point(a,n),s.calcTransformMatrix())}}class e{constructor(){this.originX=0,this.originY=0,this.drawingMode=2}make(t,i,s,e,a,n){if(this.originX=t,this.originY=i,n){const a=Math.max(Math.abs(t-s),Math.abs(i-e));s=t>s?t-a:t+a,e=i>e?i-a:i+a}return new fabric.Rect({left:s<t?s:t,top:e<i?e:i,width:Math.abs(t-s),height:Math.abs(i-e),...a})}resize(t,i,s,e,a,n){if(t instanceof fabric.Rect){if(n){const t=Math.max(Math.abs(this.originX-i),Math.abs(this.originY-s));i=this.originX>i?this.originX-t:this.originX+t,s=this.originY>s?this.originY-t:this.originY+t}t.set({originX:this.originX>i?"right":"left",originY:this.originY>s?"bottom":"top",width:Math.abs(this.originX-i),height:Math.abs(this.originY-s)}).setCoords()}return t}}class a{constructor(){this.originX=0,this.originY=0,this.drawingMode=4}make(t,i,s,e,a,n,r){this.originX=t,this.originY=i;let o=s-t,h=e-i,c=0;if(n){const t=Math.abs(o),i=Math.abs(h),s=t>=i,e=s?o:h;Math.abs(t-i)<=3?(s?(o=e,h=e*(o>=0!=h>=0?-1:1)):(o=e*(o>=0!=h>=0?-1:1),h=e),c=Math.abs(Math.sqrt(e*e*2))):s?(h=0,c=Math.abs(e)):(o=0,c=Math.abs(e))}else c=Math.abs(Math.sqrt(o*o+h*h));let d=Math.atan2(h,o);d*=180/Math.PI,d-=90;const f={left:s<t?s:t,top:e<i?e:i,fontSize:c,angle:d,fill:a.fill||(a.stroke&&"transparent"!==a.stroke?a.stroke:"black"),stroke:"transparent"!==a.stroke||a.fill&&"transparent"!==a.fill?a.stroke:"black",strokeWidth:a.strokeWidth,strokeUniform:!0,perPixelTargetFind:!0,selectable:!0};r&&(f.fontFamily=r);const l=new fabric.IText("",f);return l.controls.tqc=new fabric.Control({positionHandler:this.textCurvePositionHandler,actionHandler:this.textCurveActionHandler,actionName:"modifyTextCurve"}),l.enterEditing(),l}resize(t,i,s,e,a,n){if(t instanceof fabric.IText){let e=i-this.originX,a=s-this.originY,r=0;if(n){const t=Math.abs(e),i=Math.abs(a),s=t>=i,n=s?e:a;Math.abs(t-i)<=3?(s?(e=n,a=n*(e>=0!=a>=0?-1:1)):(e=n*(e>=0!=a>=0?-1:1),a=n),r=Math.abs(Math.sqrt(n*n*2))):s?(a=0,r=Math.abs(n)):(e=0,r=Math.abs(n))}else r=Math.abs(Math.sqrt(e*e+a*a));let o=Math.atan2(a,e);o*=180/Math.PI,o-=90,t.set({fontSize:r,angle:o}).setCoords()}return t}textCurveActionHandler(t,i,s,e){const a=i.target;if(!(c&&a instanceof fabric.IText))return!1;if(c&&c.shift)return a.curvePoint={x:0,y:0},a.path=void 0,a.curvePoint&&(0!==a.curvePoint.x||0!==a.curvePoint.y);const n=a.width,r=a._calcRotateMatrix(),o=a._calcTranslateMatrix(),h=a.getViewportTransform(),d=fabric.util.multiplyTransformMatrices(h,o);let f=fabric.util.multiplyTransformMatrices(d,r);f=fabric.util.multiplyTransformMatrices(f,[1/h[0],0,0,1/h[3],0,0]);const l=a._calculateCurrentDimensions(),b=fabric.util.transformPoint(new fabric.Point(-.5*l.x,0),f),p=fabric.util.transformPoint(new fabric.Point(.5*l.x,0),f),v=[["M",b.x,b.y],["Q",s,e,p.x,p.y]],u=new fabric.Path(v),g=fabric.util.transformPoint(new fabric.Point(s,e),fabric.util.invertTransform(f)),_={x:g.x/l.x,y:g.y/l.y};return a.curvePoint=_,c.fabricCanvas.add(u),u.set({width:n}).setCoords(),a.set({path:u}),c.fabricCanvas.remove(u),!0}textCurvePositionHandler(t,i,s){return s instanceof fabric.IText?(s.curvePoint||(s.curvePoint={x:0,y:0}),fabric.util.transformPoint(new fabric.Point(s.curvePoint.x*t.x,s.curvePoint.y*t.y),i)):new fabric.Point(0,0)}}class n{constructor(){this.originX=0,this.originY=0,this.drawingMode=5}make(t,i,s,e,a,n){if(this.originX=t,this.originY=i,n){const a=Math.max(Math.abs(t-s),Math.abs(i-e));s=t>s?t-a:t+a,e=i>e?i-a:i+a}return new fabric.Triangle({left:s<t?s:t,top:e<i?e:i,width:Math.abs(t-s),height:Math.abs(i-e),...a})}resize(t,i,s,e,a,n){if(t instanceof fabric.Triangle){if(n){const t=Math.max(Math.abs(this.originX-i),Math.abs(this.originY-s));i=this.originX>i?this.originX-t:this.originX+t,s=this.originY>s?this.originY-t:this.originY+t}t.set({originX:this.originX>i?"right":"left",originY:this.originY>s?"bottom":"top",width:Math.abs(this.originX-i),height:Math.abs(this.originY-s)}).setCoords()}return t}}class r{constructor(t){this.fabricCanvas=t,this._locked=!1,this._maxCount=100,this._redoStack=[],this._undoStack=[],this._currentState=t.toDatalessJSON()}redo(t){this.applyState(this._undoStack,this._redoStack.pop(),t)}saveState(){this._locked||(this._undoStack.length>=this._maxCount&&this._undoStack.shift(),this._undoStack.push(this._currentState),this._currentState=this.fabricCanvas.toDatalessJSON(),this._redoStack=[])}undo(t){this.applyState(this._redoStack,this._undoStack.pop(),t)}applyState(t,i,s){if(!i)return;t.push(this._currentState),this._currentState=i;const e=this;this._locked=!0,this.fabricCanvas.loadFromJSON(this._currentState,(function(){"function"==typeof s&&s(),e._locked=!1}))}}class o{constructor(t){this.editor=t}copy(t){this.editor.fabricCanvas.getActiveObject().clone((i=>{this._clipObject=i,"function"==typeof t&&t()}))}cut(){this.copy((()=>this.editor.deleteObjects()))}paste(){this._clipObject&&this._clipObject.clone((t=>{if(this.editor.fabricCanvas.discardActiveObject(),"activeSelection"===t.type){t.canvas=this.editor.fabricCanvas;let i=this;t.forEachObject((function(t){i.editor.fabricCanvas.add(t)})),t.setCoords()}else this.editor.fabricCanvas.add(t);this.editor.fabricCanvas.setActiveObject(t),this.editor.fabricCanvas.requestRenderAll(),this.editor.stateManager.saveState()}))}}class h{constructor(h){if(this._alt=!1,this._ctl=!1,this._shift=!1,this._space=!1,this._aspectRatio=4/3,this._cursorMode=0,this._enableRotation=!0,this._lastX=0,this._lastY=0,this._listeningForKeys=!0,this._mouseDown=!1,this._moving=!1,this._panning=!1,this._redraw=!1,this.editorCanvas=h,this._drawTools=[new t,new i,new e,new s,new a,new n],this._drawTool=this._drawTools[1],this._drawToolOptions={fill:"",shadow:"",stroke:"black",strokeWidth:1,perPixelTargetFind:!0,selectable:!0,strokeUniform:!0},this.fabricCanvas=new fabric.Canvas(this.editorCanvas,{selection:!1}),this.editorCanvas.parentElement&&this.editorCanvas.parentElement.parentElement){const t=this.editorCanvas.parentElement.parentElement.getBoundingClientRect();this.fabricCanvas.setDimensions({width:t.width,height:Math.round(t.width/this._aspectRatio)})}this.initializeEvents(),this.stateManager=new r(this.fabricCanvas),this.clipboard=new o(this)}get alt(){return this._alt}get ctl(){return this._ctl}get shift(){return this._shift}get space(){return this._space}bringForward(){this.fabricCanvas.getActiveObject().bringForward(!0)}bringToFront(){this.fabricCanvas.getActiveObject().bringToFront()}clear(){this.fabricCanvas.clear()}copy(){this.clipboard.copy()}cut(){this.clipboard.cut()}dispose(){window.removeEventListener("resize",this.resizeEvent),this.fabricCanvas.dispose()}deleteObjects(){this.fabricCanvas.remove(...this.fabricCanvas.getActiveObjects()),this.fabricCanvas.discardActiveObject(),this.saveState()}enableRotation(t){this._enableRotation!==t&&(this._enableRotation=t,this.fabricCanvas.forEachObject((i=>{i.setControlVisible("mtr",t)})))}paste(){this.clipboard.paste()}redo(){this.stateManager.redo()}resizeEvent(){this._resizeTimeout&&clearTimeout(this._resizeTimeout);this._resizeTimeout=setTimeout((()=>{this.editorCanvas.parentElement&&this.editorCanvas.parentElement.parentElement&&(this.fabricCanvas.setDimensions({width:"10px",height:"10px"},{cssOnly:!0}),this.resizeEventInner())}),250)}sendBackwards(){this.fabricCanvas.getActiveObject().sendBackwards(!0)}sendToBack(){this.fabricCanvas.getActiveObject().sendToBack()}setBackgroundImage(t){if(this._backgroundUrl=t,t&&this.editorCanvas.parentElement&&this.editorCanvas.parentElement.parentElement){const i=this;fabric.Image.fromURL(t,(function(t){t&&t.width&&t.height?(i._aspectRatio=t.width/t.height,i._borderWidth?(i.fabricCanvas.setDimensions({width:t.width+2*i._borderWidth,height:t.height+2*i._borderWidth},{backstoreOnly:!0}),i.fabricCanvas.setBackgroundImage(t,i.fabricCanvas.renderAll.bind(i.fabricCanvas),{top:i._borderWidth,left:i._borderWidth,originX:"left",originY:"top"})):(i.fabricCanvas.setDimensions({width:t.width,height:t.height},{backstoreOnly:!0}),i.fabricCanvas.setBackgroundImage(t,i.fabricCanvas.renderAll.bind(i.fabricCanvas))),i.resizeEvent()):i.fabricCanvas.setBackgroundImage("",i.fabricCanvas.renderAll.bind(i.fabricCanvas))}))}else this.fabricCanvas.setBackgroundImage("",this.fabricCanvas.renderAll.bind(this.fabricCanvas));this.saveState()}setBorderColor(t){this._borderColor!==t&&(this._borderColor=t,this._borderWidth&&this.drawBorder())}setBorderPercent(t){if(!t)return void this.setBorderWidth(t);const i=Math.min(this.fabricCanvas.width??0,this.fabricCanvas.height??0);i>0&&this.setBorderWidth(i*t)}setBorderWidth(t){this._borderWidth!==t&&(this._borderWidth=t,this.drawBorder(),this._backgroundUrl&&this.setBackgroundImage(this._backgroundUrl))}setDrawingMode(t){6===t?this.fabricCanvas.isDrawingMode=!0:(this.fabricCanvas.isDrawingMode=!1,this._drawTool=this._drawTools[t])}setFillColor(t){this._drawToolOptions.fill=t||"",this.setToolOptions()}setFont(t){if(!t)return;const i=this;WebFont.load({classes:!1,google:{families:[t]},active:function(){i._font=t;const s=i.fabricCanvas.getActiveObject();s instanceof fabric.IText&&s.set({fontFamily:t})}})}setLineCap(t){this._drawToolOptions.strokeLineCap=t,this.setToolOptions()}setRedraw(t){this._redraw=t}setStrokeColor(t){t?(this._drawToolOptions.stroke=t,this.fabricCanvas.freeDrawingBrush.color=t):(this._drawToolOptions.stroke="transparent",this.fabricCanvas.freeDrawingBrush.color="black"),this.setToolOptions()}setStrokeDash1(t){if(t){const i=t*(this._drawToolOptions.strokeWidth??1),s=this._drawToolOptions.strokeDashArray&&this._drawToolOptions.strokeDashArray.length&&this._drawToolOptions.strokeDashArray.length>=3?this._drawToolOptions.strokeDashArray[2]:i;let e=Math.max(this._drawToolOptions.strokeWidth??1,Math.min(i,s)/2);"round"===this._drawToolOptions.strokeLineCap&&(e+=this._drawToolOptions.strokeWidth??1),this._drawToolOptions.strokeDashArray=[i,e,s,e],this.fabricCanvas.freeDrawingBrush.strokeDashArray=[i,e,s,e]}else this._drawToolOptions.strokeDashArray=void 0,this.fabricCanvas.freeDrawingBrush.strokeDashArray=void 0;this.setToolOptions()}setStrokeDash2(t){if(t){const i=t*(this._drawToolOptions.strokeWidth??1),s=this._drawToolOptions.strokeDashArray&&this._drawToolOptions.strokeDashArray.length&&this._drawToolOptions.strokeDashArray.length>=1?this._drawToolOptions.strokeDashArray[0]:i;let e=Math.max(this._drawToolOptions.strokeWidth??1,Math.min(i,s)/2);"round"===this._drawToolOptions.strokeLineCap&&(e+=this._drawToolOptions.strokeWidth??1),this._drawToolOptions.strokeDashArray=[s,e,i,e],this.fabricCanvas.freeDrawingBrush.strokeDashArray=[s,e,i,e]}else this._drawToolOptions.strokeDashArray=void 0,this.fabricCanvas.freeDrawingBrush.strokeDashArray=void 0;this.setToolOptions()}setStrokeWidth(t){if(t??(t=1),this._drawToolOptions.strokeWidth!==t){if(this._drawToolOptions.strokeDashArray&&this._drawToolOptions.strokeDashArray.length&&this._drawToolOptions.strokeDashArray.length>=3){const i=this._drawToolOptions.strokeDashArray[0]*t/(this._drawToolOptions.strokeWidth??1),s=this._drawToolOptions.strokeDashArray[2]*t/(this._drawToolOptions.strokeWidth??1),e=Math.max(this._drawToolOptions.strokeWidth??1,Math.min(i,s)/2);this._drawToolOptions.strokeDashArray=[i,e,s,e]}this._drawToolOptions.strokeWidth=t,this.fabricCanvas.freeDrawingBrush.width=t,this.setToolOptions()}}undo(){this.stateManager.undo()}constrainView(){const t=this.fabricCanvas.viewportTransform,i=this.fabricCanvas.getZoom();if(!t)return;const s=this.fabricCanvas.getWidth();t[4]>=0?t[4]=0:t[4]<s-s*i&&(t[4]=s-s*i);const e=this.fabricCanvas.getHeight();t[5]>=0?t[5]=0:t[5]<e-e*i&&(t[5]=e-e*i)}drawBorder(){this._border&&this.fabricCanvas.remove(this._border),this._borderWidth&&(this._border=new fabric.Rect({left:0,top:0,width:this.fabricCanvas.getWidth()-this._borderWidth,height:this.fabricCanvas.getHeight()-this._borderWidth,fill:"",stroke:this._borderColor||"white",strokeWidth:this._borderWidth,selectable:!1,evented:!1}),this.fabricCanvas.add(this._border),this.fabricCanvas.requestRenderAll())}initializeEvents(){this.fabricCanvas.on("mouse:down",(t=>{const i=t.e;this._alt=i.altKey,this._ctl=i.ctrlKey,this._shift=i.shiftKey;const s=this.fabricCanvas.getPointer(t.e);this.mouseDown(s.x,s.y)})),this.fabricCanvas.on("mouse:move",(t=>{if(this._mouseDown){const i=t.e;this._alt=i.altKey,this._ctl=i.ctrlKey,this._shift=i.shiftKey;const s=this.fabricCanvas.getPointer(t.e);this.mouseMove(s.x,s.y)}})),this.fabricCanvas.on("mouse:up",(t=>{this._mouseDown=!1,this.fabricCanvas.selection=!0,this._panning&&this.fabricCanvas.viewportTransform&&this.fabricCanvas.setViewportTransform(this.fabricCanvas.viewportTransform),this._panning=!1,this._createdObject instanceof fabric.IText&&(this._cursorMode=1,this._createdObject.hiddenTextarea&&this._createdObject.hiddenTextarea.focus()),this._moving&&(this._createdObject&&this.fabricCanvas.setActiveObject(this._createdObject),this.saveState(),this._moving=!1),this._createdObject=void 0})),this.fabricCanvas.on("mouse:wheel",(t=>{const i=t.e;let s=Math.max(1,Math.min(20,this.fabricCanvas.getZoom()*.999**i.deltaY));this.fabricCanvas.zoomToPoint(new fabric.Point(i.offsetX,i.offsetY),s),i.preventDefault(),i.stopPropagation(),this.constrainView()})),this.fabricCanvas.on("object:modified",(t=>{this.saveState()})),this.fabricCanvas.on("selection:created",(t=>{this._cursorMode=1})),this.fabricCanvas.on("selection:cleared",(t=>{this._cursorMode=0})),this.fabricCanvas.on("text:editing:entered",(t=>{this._listeningForKeys=!1,this._cursorMode=1;const i=t.target;i instanceof fabric.IText&&i.curvePoint&&(0!==i.curvePoint.x||0!==i.curvePoint.y)&&(i.curvePoint={x:0,y:0},i.path=void 0)})),this.fabricCanvas.on("text:editing:exited",(t=>{this._listeningForKeys=!0,this._cursorMode=0,t.target&&this.fabricCanvas.setActiveObject(t.target),this.saveState()})),window.addEventListener("resize",this.resizeEvent.bind(this)),window.addEventListener("keydown",this.keyDownEvent.bind(this)),window.addEventListener("keyup",this.keyUpEvent.bind(this))}keyDownEvent(t){this._listeningForKeys&&(" "===t.key?this._space=!0:"Delete"===t.key?this.deleteObjects():"Copy"===t.key||"c"===t.key&&t.ctrlKey?this.copy():"Cut"===t.key||"x"===t.key&&t.ctrlKey?this.cut():"Paste"===t.key||"v"===t.key&&t.ctrlKey?this.paste():"Undo"===t.key||"z"===t.key&&t.ctrlKey?this.undo():("Redo"===t.key||"y"===t.key&&t.ctrlKey)&&this.redo())}keyUpEvent(t){" "===t.key&&(this._space=!1)}make(t,i,s,e){this._redraw&&this.fabricCanvas.remove(...this.fabricCanvas.getObjects());const a=this._drawTool.make(t,i,s,e,this._drawToolOptions,this._shift,this._font);return this._enableRotation||a.setControlVisible("mtr",!1),this.fabricCanvas.add(a),this.saveState(),a}mouseDown(t,i){if(this._space)this._mouseDown=!0,this._panning=!0,this.fabricCanvas.selection=!1,this._lastX=t,this._lastY=i;else if(this._alt){const s=this.fabricCanvas.getActiveObject();if(s instanceof fabric.Path){const e=this._drawTools[3];this._mouseDown=!0,this.fabricCanvas.selection=!1,e.addPoint(s,t,i)}}else 0!==this._cursorMode||this.fabricCanvas.isDrawingMode||this._alt||this._ctl||0!==this.fabricCanvas.getActiveObjects().length||(this._mouseDown=!0,this.fabricCanvas.selection=!1,this._lastX=t,this._lastY=i)}mouseMove(t,i){if(this._panning){const s=this.fabricCanvas.viewportTransform,e=this.fabricCanvas.getZoom();s&&(s[4]+=(t-this._lastX)*e,s[5]+=(i-this._lastY)*e,this.constrainView(),this.fabricCanvas.requestRenderAll()),this._lastX=t,this._lastY=i}else this._createdObject?(this._moving=!0,this._drawTool.resize(this._createdObject,t,i,this._ctl,this._alt,this._shift),this.fabricCanvas.renderAll()):this._createdObject=this.make(this._lastX,this._lastY,t,i)}resizeEventInner(){this._resizeTimeoutInner&&clearTimeout(this._resizeTimeoutInner);this._resizeTimeoutInner=setTimeout((()=>{if(this.editorCanvas.parentElement&&this.editorCanvas.parentElement.parentElement){const t=this.editorCanvas.parentElement.parentElement.getBoundingClientRect();this.fabricCanvas.setDimensions({width:`${t.width}px`,height:`${Math.round(t.width/this._aspectRatio)}px`},{cssOnly:!0}),this.drawBorder()}}),10)}saveState(){this.stateManager.saveState(),this.fabricCanvas.renderAll()}setToolOptions(){this.fabricCanvas.getActiveObjects().forEach((t=>{t.set(this._drawToolOptions)})),this.fabricCanvas.requestRenderAll()}}let c;function d(){c&&c.bringForward()}function f(){c&&c.bringToFront()}function l(){c&&c.clear()}function b(){c&&c.copy()}function p(){c&&c.cut()}function v(){c&&c.deleteObjects()}function u(){c&&c.dispose()}function g(t){c&&c.enableRotation(t||!1)}function _(){return c?JSON.stringify(c.fabricCanvas):""}async function w(){if(!c)return"";return await new Promise((function(t){c?(c.fabricCanvas.discardActiveObject().renderAll(),c.editorCanvas.toBlob((i=>{t(i?URL.createObjectURL(i):"")}))):t("")}))}function C(t,i){c?c.clear():c=new h(t),i&&c.setBackgroundImage(i)}function m(t){c&&t&&c.fabricCanvas.loadFromJSON(t,(()=>{const t=[];c.fabricCanvas.forEachObject((i=>{i instanceof fabric.IText&&i.fontFamily&&t.push(i.fontFamily)})),t.length&&WebFont.load({classes:!1,google:{families:t},active:function(){c.fabricCanvas.requestRenderAll()}})}))}function y(){c&&c.paste()}function k(){c&&c.redo()}function x(){c&&c.resizeEvent()}function M(t){URL.revokeObjectURL(t)}function O(){if(c){const t=document.createElement("a");t.setAttribute("download","image.png"),t.setAttribute("href",c.fabricCanvas.toDataURL({format:"image/png"}).replace("image/png","image/octet-stream")),t.click()}}function T(t){c&&c.setBackgroundImage(t)}function P(){c&&c.sendBackwards()}function D(){c&&c.sendToBack()}function I(t){c&&c.setBorderColor(t)}function A(t){c&&c.setBorderPercent(t)}function S(t){c&&c.setBorderWidth(t)}function E(t){c&&c.setDrawingMode(t)}function j(t){c&&c.setFillColor(t)}function Y(t){t&&c&&c.setFont(t)}function B(t){c&&c.setLineCap(t)}function X(t){c&&c.setRedraw(t||!1)}function W(t){c&&c.setStrokeColor(t)}function R(t){c&&c.setStrokeDash1(t)}function z(t){c&&c.setStrokeDash2(t)}function F(t){c&&c.setStrokeWidth(t)}function H(){c&&c.undo()}export{d as bringForward,f as bringToFront,l as clear,b as copy,p as cut,v as deleteObjects,u as dispose,g as enableRotation,_ as getJSON,w as getObjUrl,C as loadEditor,m as loadJSON,y as paste,k as redo,x as resize,M as revokeObjUrl,O as saveImage,P as sendBackwards,D as sendToBack,T as setBackgroundImage,I as setBorderColor,A as setBorderPercent,S as setBorderWidth,E as setDrawingMode,j as setFillColor,Y as setFont,B as setLineCap,X as setRedraw,W as setStrokeColor,R as setStrokeDash1,z as setStrokeDash2,F as setStrokeWidth,H as undo};
+class t{constructor(){this.originX=0,this.originY=0,this.drawingMode=0}make(t,i,s,e,a,n){if(this.originX=t,this.originY=i,n){const a=Math.max(Math.abs(t-s),Math.abs(i-e));s=t>s?t-a:t+a,e=i>e?i-a:i+a}return new fabric.Ellipse({left:s<t?s:t,top:e<i?e:i,rx:Math.abs(s-t)/2,ry:Math.abs(s-t)/2,...a})}resize(t,i,s,e,a,n){if(t instanceof fabric.Ellipse){if(n){const t=Math.max(Math.abs(this.originX-i),Math.abs(this.originY-s));i=this.originX>i?this.originX-t:this.originX+t,s=this.originY>s?this.originY-t:this.originY+t}t.set({originX:this.originX>i?"right":"left",originY:this.originY>s?"bottom":"top",rx:Math.abs(this.originX-i)/2,ry:Math.abs(this.originY-s)/2}).setCoords()}return t}}class i{constructor(){this.drawingMode=1}make(t,i,s,e,a,n){if(n){const a=Math.abs(t-s),n=Math.abs(i-e),r=Math.max(a,n);Math.abs(a-n)<=3?(s=t>s?t-r:t+r,e=i>e?i-r:i+r):a>n?(s=t>s?t-r:t+r,e=i):(s=t,e=i>e?i-r:i+r)}return new fabric.Line([t,i,s,e],a)}resize(t,i,s,e,a,n){if(t instanceof fabric.Line){if(n){const e=Math.abs((t.x1??0)-i),a=Math.abs((t.y1??0)-s),n=Math.max(e,a);Math.abs(e-a)<=3?(i=(t.x1??0)>i?(t.x1??0)-n:(t.x1??0)+n,s=(t.y1??0)>s?(t.y1??0)-n:(t.y1??0)+n):e>a?(i=(t.x1??0)>i?(t.x1??0)-n:(t.x1??0)+n,s=t.y1??0):(i=t.x1??0,s=(t.y1??0)>s?(t.y1??0)-n:(t.y1??0)+n)}t.set({x2:i,y2:s}).setCoords()}return t}}class s{constructor(){this.drawingMode=1}addPoint(t,i,s){const e=t.path,a=e.length+1;e.push(["Q",i,s,i,s]);const n=t.getCenterPoint(),r=new fabric.Path(e);t.set({path:r.path,width:r.width,height:r.height,dirty:!0,pathOffset:r.pathOffset}).setCoords(),t.setPositionByOrigin(n,"center","center"),this.addControlPoint(t,a-1)}make(t,i,s,e,a,n){const r=[[],[]];if(n){const a=Math.abs(t-s),n=Math.abs(i-e),r=Math.max(a,n);Math.abs(a-n)<=3?(s=t>s?t-r:t+r,e=i>e?i-r:i+r):a>n?(s=t>s?t-r:t+r,e=i):(s=t,e=i>e?i-r:i+r)}r[0][0]="M",r[0][1]=t,r[0][2]=i,r[1][0]="Q",r[1][1]=s,r[1][2]=e,r[1][3]=s,r[1][4]=e;const o=new fabric.Path(r,a);return o.controls={...fabric.Object.prototype.controls},o.controls.csc=new fabric.Control({positionHandler:this.curvePositionHandler,actionHandler:this.curveActionHandler,actionName:"modifyCurve"}),o.controls.csc.pointIndex=0,this.addControlPoint(o,1),o}resize(t,i,s,e,a,n){if(t instanceof fabric.Path&&t.path){const e=t.path,a=e.length-1;if(n&&1===a){const t=Math.abs(e[0][1]-i),a=Math.abs(e[0][2]-s),n=Math.max(t,a);Math.abs(t-a)<=3?(i=e[0][1]>i?e[0][1]-n:e[0][1]+n,s=e[0][2]>s?e[0][2]-n:e[0][2]+n):t>a?(i=e[0][1]>i?e[0][1]-n:e[0][1]+n,s=e[0][2]):(i=e[0][1],s=e[0][2]>s?e[0][2]-n:e[0][2]+n)}const r=e;r[a][1]=i,r[a][2]=s,r[a][3]=i,r[a][4]=s;const o=t.getCenterPoint(),h=new fabric.Path(r);t.set({path:h.path,width:h.width,height:h.height,dirty:!0,pathOffset:h.pathOffset}),1===a&&t.set({left:i<e[0][1]?i:e[0][1],top:s<e[0][2]?s:e[0][2]}),t.setCoords(),1!==a&&t.setPositionByOrigin(o,"center","center")}return t}addControlPoint(t,i){t.controls[`cqc${i}`]=new fabric.Control({positionHandler:this.curvePositionHandler,actionHandler:this.curveActionHandler,actionName:"modifyCurve"});let s=t.controls[`cqc${i}`];s.pointIndex=i,s.pointIsCurve=!0,t.controls[`cpc${i}`]=new fabric.Control({positionHandler:this.curvePositionHandler,actionHandler:this.curveActionHandler,actionName:"modifyCurve"}),s=t.controls[`cpc${i}`],s.pointIndex=i}curveActionHandler(t,i,s,e){if(!(i.target instanceof fabric.Path))return!1;const a=i.target,n=a.path,r=a.controls[a.__corner],o=a._getNonTransformedDimensions(),h=a._getTransformedDimensions(),d=a.strokeWidth??1,l=a.strokeUniform?d:d*(a.scaleX??1),f=(a.strokeUniform||a.scaleY,(o.x-d)/(h.x-l)),b=(o.y-d)/(h.y-l),p=a.toLocalPoint(new fabric.Point(s,e),"center","center"),v=new fabric.Point(p.x*f+a.pathOffset.x,p.y*b+a.pathOffset.y);0===r.pointIndex?(n[0][1]=v.x,n[0][2]=v.y):r.pointIsCurve?c&&c.shift?(n[r.pointIndex][1]=n[r.pointIndex][3],n[r.pointIndex][2]=n[r.pointIndex][4]):(n[r.pointIndex][1]=v.x,n[r.pointIndex][2]=v.y):(n[r.pointIndex][1]===n[r.pointIndex][3]&&n[r.pointIndex][2]===n[r.pointIndex][4]&&(n[r.pointIndex][1]=v.x,n[r.pointIndex][2]=v.y),n[r.pointIndex][3]=v.x,n[r.pointIndex][4]=v.y);const g=a.getCenterPoint(),u=new fabric.Path(n);return a.set({path:u.path,width:u.width,height:u.height,dirty:!0,pathOffset:u.pathOffset}).setCoords(),a.setPositionByOrigin(g,"center","center"),!0}curvePositionHandler(t,i,s){if(!(s instanceof fabric.Path))return new fabric.Point(0,0);const e=s.path;let a,n;if(0===this.pointIndex)a=e[0][1]-s.pathOffset.x,n=e[0][2]-s.pathOffset.y;else if(this.pointIsCurve)if(e[this.pointIndex][1]===e[this.pointIndex][3]&&e[this.pointIndex][2]===e[this.pointIndex][4]){const t=1===this.pointIndex?1:3,i=1===this.pointIndex?2:4;a=(e[this.pointIndex-1][t]+e[this.pointIndex][3])/2-s.pathOffset.x,n=(e[this.pointIndex-1][i]+e[this.pointIndex][4])/2-s.pathOffset.y}else a=e[this.pointIndex][1]-s.pathOffset.x,n=e[this.pointIndex][2]-s.pathOffset.y;else a=e[this.pointIndex][3]-s.pathOffset.x,n=e[this.pointIndex][4]-s.pathOffset.y;return s.canvas&&s.canvas.viewportTransform?fabric.util.transformPoint(new fabric.Point(a,n),fabric.util.multiplyTransformMatrices(s.canvas.viewportTransform,s.calcTransformMatrix())):fabric.util.transformPoint(new fabric.Point(a,n),s.calcTransformMatrix())}}class e{constructor(){this.originX=0,this.originY=0,this.drawingMode=2}make(t,i,s,e,a,n){if(this.originX=t,this.originY=i,n){const a=Math.max(Math.abs(t-s),Math.abs(i-e));s=t>s?t-a:t+a,e=i>e?i-a:i+a}return new fabric.Rect({left:s<t?s:t,top:e<i?e:i,width:Math.abs(t-s),height:Math.abs(i-e),...a})}resize(t,i,s,e,a,n){if(t instanceof fabric.Rect){if(n){const t=Math.max(Math.abs(this.originX-i),Math.abs(this.originY-s));i=this.originX>i?this.originX-t:this.originX+t,s=this.originY>s?this.originY-t:this.originY+t}t.set({originX:this.originX>i?"right":"left",originY:this.originY>s?"bottom":"top",width:Math.abs(this.originX-i),height:Math.abs(this.originY-s)}).setCoords()}return t}}class a{constructor(){this.originX=0,this.originY=0,this.drawingMode=4}make(t,i,s,e,a,n,r){this.originX=t,this.originY=i;let o=s-t,h=e-i,c=0;if(n){const t=Math.abs(o),i=Math.abs(h),s=t>=i,e=s?o:h;Math.abs(t-i)<=3?(s?(o=e,h=e*(o>=0!=h>=0?-1:1)):(o=e*(o>=0!=h>=0?-1:1),h=e),c=Math.abs(Math.sqrt(e*e*2))):s?(h=0,c=Math.abs(e)):(o=0,c=Math.abs(e))}else c=Math.abs(Math.sqrt(o*o+h*h));let d=Math.atan2(h,o);d*=180/Math.PI,d-=90;const l={left:s<t?s:t,top:e<i?e:i,fontSize:c,angle:d,fill:a.fill||(a.stroke&&"transparent"!==a.stroke?a.stroke:"black"),stroke:"transparent"!==a.stroke||a.fill&&"transparent"!==a.fill?a.stroke:"black",strokeWidth:a.strokeWidth,strokeUniform:!0,perPixelTargetFind:!0,selectable:!0};r&&(l.fontFamily=r);const f=new fabric.IText("",l);return f.controls.tqc=new fabric.Control({positionHandler:this.textCurvePositionHandler,actionHandler:this.textCurveActionHandler,actionName:"modifyTextCurve"}),f.enterEditing(),f}resize(t,i,s,e,a,n){if(t instanceof fabric.IText){let e=i-this.originX,a=s-this.originY,r=0;if(n){const t=Math.abs(e),i=Math.abs(a),s=t>=i,n=s?e:a;Math.abs(t-i)<=3?(s?(e=n,a=n*(e>=0!=a>=0?-1:1)):(e=n*(e>=0!=a>=0?-1:1),a=n),r=Math.abs(Math.sqrt(n*n*2))):s?(a=0,r=Math.abs(n)):(e=0,r=Math.abs(n))}else r=Math.abs(Math.sqrt(e*e+a*a));let o=Math.atan2(a,e);o*=180/Math.PI,o-=90,t.set({fontSize:r,angle:o}).setCoords()}return t}textCurveActionHandler(t,i,s,e){const a=i.target;if(!(c&&a instanceof fabric.IText))return!1;if(c&&c.shift)return a.curvePoint={x:0,y:0},a.path=void 0,a.curvePoint&&(0!==a.curvePoint.x||0!==a.curvePoint.y);const n=a.width,r=a._calcRotateMatrix(),o=a._calcTranslateMatrix(),h=a.getViewportTransform(),d=fabric.util.multiplyTransformMatrices(h,o);let l=fabric.util.multiplyTransformMatrices(d,r);l=fabric.util.multiplyTransformMatrices(l,[1/h[0],0,0,1/h[3],0,0]);const f=a._calculateCurrentDimensions(),b=fabric.util.transformPoint(new fabric.Point(-.5*f.x,0),l),p=fabric.util.transformPoint(new fabric.Point(.5*f.x,0),l),v=[["M",b.x,b.y],["Q",s,e,p.x,p.y]],g=new fabric.Path(v),u=fabric.util.transformPoint(new fabric.Point(s,e),fabric.util.invertTransform(l)),_={x:u.x/f.x,y:u.y/f.y};return a.curvePoint=_,c.fabricCanvas.add(g),g.set({width:n}).setCoords(),a.set({path:g}),c.fabricCanvas.remove(g),!0}textCurvePositionHandler(t,i,s){return s instanceof fabric.IText?(s.curvePoint||(s.curvePoint={x:0,y:0}),fabric.util.transformPoint(new fabric.Point(s.curvePoint.x*t.x,s.curvePoint.y*t.y),i)):new fabric.Point(0,0)}}class n{constructor(){this.originX=0,this.originY=0,this.drawingMode=5}make(t,i,s,e,a,n){if(this.originX=t,this.originY=i,n){const a=Math.max(Math.abs(t-s),Math.abs(i-e));s=t>s?t-a:t+a,e=i>e?i-a:i+a}return new fabric.Triangle({left:s<t?s:t,top:e<i?e:i,width:Math.abs(t-s),height:Math.abs(i-e),...a})}resize(t,i,s,e,a,n){if(t instanceof fabric.Triangle){if(n){const t=Math.max(Math.abs(this.originX-i),Math.abs(this.originY-s));i=this.originX>i?this.originX-t:this.originX+t,s=this.originY>s?this.originY-t:this.originY+t}t.set({originX:this.originX>i?"right":"left",originY:this.originY>s?"bottom":"top",width:Math.abs(this.originX-i),height:Math.abs(this.originY-s)}).setCoords()}return t}}class r{constructor(t){this.fabricCanvas=t,this._locked=!1,this._maxCount=100,this._redoStack=[],this._undoStack=[],this._currentState=t.toDatalessJSON()}redo(t){this.applyState(this._undoStack,this._redoStack.pop(),t)}saveState(){this._locked||(this._undoStack.length>=this._maxCount&&this._undoStack.shift(),this._undoStack.push(this._currentState),this._currentState=this.fabricCanvas.toDatalessJSON(),this._redoStack=[])}undo(t){this.applyState(this._redoStack,this._undoStack.pop(),t)}applyState(t,i,s){if(!i)return;t.push(this._currentState),this._currentState=i;const e=this;this._locked=!0,this.fabricCanvas.loadFromJSON(this._currentState,(function(){"function"==typeof s&&s(),e._locked=!1}))}}class o{constructor(t){this.editor=t}copy(t){this.editor.fabricCanvas.getActiveObject().clone((i=>{this._clipObject=i,"function"==typeof t&&t()}))}cut(){this.copy((()=>this.editor.deleteObjects()))}paste(){this._clipObject&&this._clipObject.clone((t=>{if(this.editor.fabricCanvas.discardActiveObject(),"activeSelection"===t.type){t.canvas=this.editor.fabricCanvas;let i=this;t.forEachObject((function(t){i.editor.fabricCanvas.add(t)})),t.setCoords()}else this.editor.fabricCanvas.add(t);this.editor.fabricCanvas.setActiveObject(t),this.editor.fabricCanvas.requestRenderAll(),this.editor.stateManager.saveState()}))}}class h{constructor(h){if(this._alt=!1,this._ctl=!1,this._shift=!1,this._space=!1,this._aspectRatio=4/3,this._cursorMode=0,this._enableRotation=!0,this._lastX=0,this._lastY=0,this._listeningForKeys=!0,this._mouseDown=!1,this._moving=!1,this._panning=!1,this._redraw=!1,this.editorCanvas=h,this._drawTools=[new t,new i,new e,new s,new a,new n],this._drawTool=this._drawTools[1],this._drawToolOptions={fill:"",shadow:"",stroke:"black",strokeWidth:1,perPixelTargetFind:!0,selectable:!0,strokeUniform:!0},this.fabricCanvas=new fabric.Canvas(this.editorCanvas,{selection:!1}),this.editorCanvas.parentElement&&this.editorCanvas.parentElement.parentElement){const t=this.editorCanvas.parentElement.parentElement.getBoundingClientRect();this.fabricCanvas.setDimensions({width:t.width,height:Math.round(t.width/this._aspectRatio)})}this.initializeEvents(),this.stateManager=new r(this.fabricCanvas),this.clipboard=new o(this)}get alt(){return this._alt}get ctl(){return this._ctl}get shift(){return this._shift}get space(){return this._space}bringForward(){this.fabricCanvas.getActiveObject().bringForward(!0)}bringToFront(){this.fabricCanvas.getActiveObject().bringToFront()}clear(){this.fabricCanvas.clear()}copy(){this.clipboard.copy()}cut(){this.clipboard.cut()}dispose(){window.removeEventListener("resize",this.resizeEvent),this.fabricCanvas.dispose()}deleteObjects(){this.fabricCanvas.remove(...this.fabricCanvas.getActiveObjects()),this.fabricCanvas.discardActiveObject(),this.saveState()}enableRotation(t){this._enableRotation!==t&&(this._enableRotation=t,this.fabricCanvas.forEachObject((i=>{i.setControlVisible("mtr",t)})))}paste(){this.clipboard.paste()}redo(){this.stateManager.redo()}resizeEvent(){this._resizeTimeout&&clearTimeout(this._resizeTimeout);this._resizeTimeout=setTimeout((()=>{this.editorCanvas.parentElement&&this.editorCanvas.parentElement.parentElement&&(this.fabricCanvas.setDimensions({width:"10px",height:"10px"},{cssOnly:!0}),this.resizeEventInner())}),250)}sendBackwards(){this.fabricCanvas.getActiveObject().sendBackwards(!0)}sendToBack(){this.fabricCanvas.getActiveObject().sendToBack()}setBackgroundImage(t){if(this._backgroundUrl=t,t&&this.editorCanvas.parentElement&&this.editorCanvas.parentElement.parentElement){const i=this;fabric.Image.fromURL(t,(function(t){t&&t.width&&t.height?(i._aspectRatio=t.width/t.height,i._borderWidth?(i.fabricCanvas.setDimensions({width:t.width+2*i._borderWidth,height:t.height+2*i._borderWidth},{backstoreOnly:!0}),i.fabricCanvas.setBackgroundImage(t,i.fabricCanvas.renderAll.bind(i.fabricCanvas),{top:i._borderWidth,left:i._borderWidth,originX:"left",originY:"top"})):(i.fabricCanvas.setDimensions({width:t.width,height:t.height},{backstoreOnly:!0}),i.fabricCanvas.setBackgroundImage(t,i.fabricCanvas.renderAll.bind(i.fabricCanvas))),i.resizeEvent()):i.fabricCanvas.setBackgroundImage("",i.fabricCanvas.renderAll.bind(i.fabricCanvas))}))}else this.fabricCanvas.setBackgroundImage("",this.fabricCanvas.renderAll.bind(this.fabricCanvas));this.saveState()}setBorderColor(t){this._borderColor!==t&&(this._borderColor=t,this._borderWidth&&this.drawBorder())}setBorderPercent(t){if(!t)return void this.setBorderWidth(t);const i=Math.min(this.fabricCanvas.width??0,this.fabricCanvas.height??0);i>0&&this.setBorderWidth(i*t)}setBorderWidth(t){this._borderWidth!==t&&(this._borderWidth=t,this.drawBorder(),this._backgroundUrl&&this.setBackgroundImage(this._backgroundUrl))}setDrawingMode(t){6===t?this.fabricCanvas.isDrawingMode=!0:(this.fabricCanvas.isDrawingMode=!1,this._drawTool=this._drawTools[t])}setFillColor(t){this._drawToolOptions.fill=t||"",this.setToolOptions()}setFont(t){if(!t)return;const i=this;WebFont.load({classes:!1,google:{families:[t]},active:function(){i._font=t;const s=i.fabricCanvas.getActiveObject();s instanceof fabric.IText&&s.set({fontFamily:t})}})}setLineCap(t){this._drawToolOptions.strokeLineCap=t,this.setToolOptions()}setRedraw(t){this._redraw=t}setStrokeColor(t){t?(this._drawToolOptions.stroke=t,this.fabricCanvas.freeDrawingBrush.color=t):(this._drawToolOptions.stroke="transparent",this.fabricCanvas.freeDrawingBrush.color="black"),this.setToolOptions()}setStrokeDash1(t){if(t){const i=t*(this._drawToolOptions.strokeWidth??1),s=this._drawToolOptions.strokeDashArray&&this._drawToolOptions.strokeDashArray.length&&this._drawToolOptions.strokeDashArray.length>=3?this._drawToolOptions.strokeDashArray[2]:i;let e=Math.max(this._drawToolOptions.strokeWidth??1,Math.min(i,s)/2);"round"===this._drawToolOptions.strokeLineCap&&(e+=this._drawToolOptions.strokeWidth??1),this._drawToolOptions.strokeDashArray=[i,e,s,e],this.fabricCanvas.freeDrawingBrush.strokeDashArray=[i,e,s,e]}else this._drawToolOptions.strokeDashArray=void 0,this.fabricCanvas.freeDrawingBrush.strokeDashArray=void 0;this.setToolOptions()}setStrokeDash2(t){if(t){const i=t*(this._drawToolOptions.strokeWidth??1),s=this._drawToolOptions.strokeDashArray&&this._drawToolOptions.strokeDashArray.length&&this._drawToolOptions.strokeDashArray.length>=1?this._drawToolOptions.strokeDashArray[0]:i;let e=Math.max(this._drawToolOptions.strokeWidth??1,Math.min(i,s)/2);"round"===this._drawToolOptions.strokeLineCap&&(e+=this._drawToolOptions.strokeWidth??1),this._drawToolOptions.strokeDashArray=[s,e,i,e],this.fabricCanvas.freeDrawingBrush.strokeDashArray=[s,e,i,e]}else this._drawToolOptions.strokeDashArray=void 0,this.fabricCanvas.freeDrawingBrush.strokeDashArray=void 0;this.setToolOptions()}setStrokeWidth(t){if(t??(t=1),this._drawToolOptions.strokeWidth!==t){if(this._drawToolOptions.strokeDashArray&&this._drawToolOptions.strokeDashArray.length&&this._drawToolOptions.strokeDashArray.length>=3){const i=this._drawToolOptions.strokeDashArray[0]*t/(this._drawToolOptions.strokeWidth??1),s=this._drawToolOptions.strokeDashArray[2]*t/(this._drawToolOptions.strokeWidth??1),e=Math.max(this._drawToolOptions.strokeWidth??1,Math.min(i,s)/2);this._drawToolOptions.strokeDashArray=[i,e,s,e]}this._drawToolOptions.strokeWidth=t,this.fabricCanvas.freeDrawingBrush.width=t,this.setToolOptions()}}undo(){this.stateManager.undo()}constrainView(){const t=this.fabricCanvas.viewportTransform,i=this.fabricCanvas.getZoom();if(!t)return;const s=this.fabricCanvas.getWidth();t[4]>=0?t[4]=0:t[4]<s-s*i&&(t[4]=s-s*i);const e=this.fabricCanvas.getHeight();t[5]>=0?t[5]=0:t[5]<e-e*i&&(t[5]=e-e*i)}drawBorder(){this._border&&this.fabricCanvas.remove(this._border),this._borderWidth&&(this._border=new fabric.Rect({left:0,top:0,width:this.fabricCanvas.getWidth()-this._borderWidth,height:this.fabricCanvas.getHeight()-this._borderWidth,fill:"",stroke:this._borderColor||"white",strokeWidth:this._borderWidth,selectable:!1,evented:!1}),this.fabricCanvas.add(this._border),this.fabricCanvas.requestRenderAll())}initializeEvents(){this.fabricCanvas.on("mouse:down",(t=>{const i=t.e;this._alt=i.altKey,this._ctl=i.ctrlKey,this._shift=i.shiftKey;const s=this.fabricCanvas.getPointer(t.e);this.mouseDown(s.x,s.y)})),this.fabricCanvas.on("mouse:move",(t=>{if(this._mouseDown){const i=t.e;this._alt=i.altKey,this._ctl=i.ctrlKey,this._shift=i.shiftKey;const s=this.fabricCanvas.getPointer(t.e);this.mouseMove(s.x,s.y)}})),this.fabricCanvas.on("mouse:up",(t=>{this._mouseDown=!1,this.fabricCanvas.selection=!0,this._panning&&this.fabricCanvas.viewportTransform&&this.fabricCanvas.setViewportTransform(this.fabricCanvas.viewportTransform),this._panning=!1,this._createdObject instanceof fabric.IText&&(this._cursorMode=1,this._createdObject.hiddenTextarea&&this._createdObject.hiddenTextarea.focus()),this._moving&&(this._createdObject&&this.fabricCanvas.setActiveObject(this._createdObject),this.saveState(),this._moving=!1),this._createdObject=void 0})),this.fabricCanvas.on("mouse:wheel",(t=>{const i=t.e;let s=Math.max(1,Math.min(20,this.fabricCanvas.getZoom()*.999**i.deltaY));this.fabricCanvas.zoomToPoint(new fabric.Point(i.offsetX,i.offsetY),s),i.preventDefault(),i.stopPropagation(),this.constrainView()})),this.fabricCanvas.on("object:modified",(t=>{this.saveState()})),this.fabricCanvas.on("selection:created",(t=>{this._cursorMode=1})),this.fabricCanvas.on("selection:cleared",(t=>{this._cursorMode=0})),this.fabricCanvas.on("text:editing:entered",(t=>{this._listeningForKeys=!1,this._cursorMode=1;const i=t.target;i instanceof fabric.IText&&i.curvePoint&&(0!==i.curvePoint.x||0!==i.curvePoint.y)&&(i.curvePoint={x:0,y:0},i.path=void 0)})),this.fabricCanvas.on("text:editing:exited",(t=>{this._listeningForKeys=!0,this._cursorMode=0,t.target&&this.fabricCanvas.setActiveObject(t.target),this.saveState()})),window.addEventListener("resize",this.resizeEvent.bind(this)),window.addEventListener("keydown",this.keyDownEvent.bind(this)),window.addEventListener("keyup",this.keyUpEvent.bind(this))}keyDownEvent(t){this._listeningForKeys&&(" "===t.key?this._space=!0:"Delete"===t.key?this.deleteObjects():"Copy"===t.key||"c"===t.key&&t.ctrlKey?this.copy():"Cut"===t.key||"x"===t.key&&t.ctrlKey?this.cut():"Paste"===t.key||"v"===t.key&&t.ctrlKey?this.paste():"Undo"===t.key||"z"===t.key&&t.ctrlKey?this.undo():("Redo"===t.key||"y"===t.key&&t.ctrlKey)&&this.redo())}keyUpEvent(t){" "===t.key&&(this._space=!1)}make(t,i,s,e){this._redraw&&this.fabricCanvas.remove(...this.fabricCanvas.getObjects());const a=this._drawTool.make(t,i,s,e,this._drawToolOptions,this._shift,this._font);return this._enableRotation||a.setControlVisible("mtr",!1),this.fabricCanvas.add(a),this.saveState(),a}mouseDown(t,i){if(this._space)this._mouseDown=!0,this._panning=!0,this.fabricCanvas.selection=!1,this._lastX=t,this._lastY=i;else if(this._alt){const s=this.fabricCanvas.getActiveObject();if(console.log(s),s instanceof fabric.Path){const e=this._drawTools[3];this._mouseDown=!0,this.fabricCanvas.selection=!1,e.addPoint(s,t,i),this.fabricCanvas.renderAll()}}else 0!==this._cursorMode||this.fabricCanvas.isDrawingMode||this._alt||this._ctl||0!==this.fabricCanvas.getActiveObjects().length||(this._mouseDown=!0,this.fabricCanvas.selection=!1,this._lastX=t,this._lastY=i)}mouseMove(t,i){if(this._panning){const s=this.fabricCanvas.viewportTransform,e=this.fabricCanvas.getZoom();s&&(s[4]+=(t-this._lastX)*e,s[5]+=(i-this._lastY)*e,this.constrainView(),this.fabricCanvas.requestRenderAll()),this._lastX=t,this._lastY=i}else this._createdObject?(this._moving=!0,this._drawTool.resize(this._createdObject,t,i,this._ctl,this._alt,this._shift),this.fabricCanvas.renderAll()):this._createdObject=this.make(this._lastX,this._lastY,t,i)}resizeEventInner(){this._resizeTimeoutInner&&clearTimeout(this._resizeTimeoutInner);this._resizeTimeoutInner=setTimeout((()=>{if(this.editorCanvas.parentElement&&this.editorCanvas.parentElement.parentElement){const t=this.editorCanvas.parentElement.parentElement.getBoundingClientRect();this.fabricCanvas.setDimensions({width:`${t.width}px`,height:`${Math.round(t.width/this._aspectRatio)}px`},{cssOnly:!0}),this.drawBorder()}}),10)}saveState(){this.stateManager.saveState(),this.fabricCanvas.renderAll()}setToolOptions(){this.fabricCanvas.getActiveObjects().forEach((t=>{t.set(this._drawToolOptions)})),this.fabricCanvas.requestRenderAll()}}let c;function d(){c&&c.bringForward()}function l(){c&&c.bringToFront()}function f(){c&&c.clear()}function b(){c&&c.copy()}function p(){c&&c.cut()}function v(){c&&c.deleteObjects()}function g(){c&&c.dispose()}function u(t){c&&c.enableRotation(t||!1)}function _(){return c?JSON.stringify(c.fabricCanvas):""}async function w(){if(!c)return"";return await new Promise((function(t){c?(c.fabricCanvas.discardActiveObject().renderAll(),c.editorCanvas.toBlob((i=>{t(i?URL.createObjectURL(i):"")}))):t("")}))}function C(t,i){c?c.clear():c=new h(t),i&&c.setBackgroundImage(i)}function m(t){c&&t&&c.fabricCanvas.loadFromJSON(t,(()=>{const t=[];c.fabricCanvas.forEachObject((i=>{i instanceof fabric.IText&&i.fontFamily&&t.push(i.fontFamily)})),t.length&&WebFont.load({classes:!1,google:{families:t},active:function(){c.fabricCanvas.requestRenderAll()}})}))}function y(){c&&c.paste()}function k(){c&&c.redo()}function x(){c&&c.resizeEvent()}function M(t){URL.revokeObjectURL(t)}function O(){if(c){const t=document.createElement("a");t.setAttribute("download","image.png"),t.setAttribute("href",c.fabricCanvas.toDataURL({format:"image/png"}).replace("image/png","image/octet-stream")),t.click()}}function T(t){c&&c.setBackgroundImage(t)}function P(){c&&c.sendBackwards()}function D(){c&&c.sendToBack()}function I(t){c&&c.setBorderColor(t)}function A(t){c&&c.setBorderPercent(t)}function S(t){c&&c.setBorderWidth(t)}function E(t){c&&c.setDrawingMode(t)}function j(t){c&&c.setFillColor(t)}function Y(t){t&&c&&c.setFont(t)}function B(t){c&&c.setLineCap(t)}function X(t){c&&c.setRedraw(t||!1)}function W(t){c&&c.setStrokeColor(t)}function R(t){c&&c.setStrokeDash1(t)}function z(t){c&&c.setStrokeDash2(t)}function F(t){c&&c.setStrokeWidth(t)}function H(){c&&c.undo()}export{d as bringForward,l as bringToFront,f as clear,b as copy,p as cut,v as deleteObjects,g as dispose,u as enableRotation,_ as getJSON,w as getObjUrl,C as loadEditor,m as loadJSON,y as paste,k as redo,x as resize,M as revokeObjUrl,O as saveImage,P as sendBackwards,D as sendToBack,T as setBackgroundImage,I as setBorderColor,A as setBorderPercent,S as setBorderWidth,E as setDrawingMode,j as setFillColor,Y as setFont,B as setLineCap,X as setRedraw,W as setStrokeColor,R as setStrokeDash1,z as setStrokeDash2,F as setStrokeWidth,H as undo};
 //# sourceMappingURL=editor.js.map
